Ticket #3032: 'External panelize' results cleared when opening another result.
The panel containing 'external panelize' results is cleared when
opening a result, then opening another.

How to reproduce:
1. C-x !
2. Launch an 'External panelize' command, either by recalling
a previously stored command or typing a new one, like 'find . -type f'.
3. When the results come up, open one by hitting the RETURN key
if a program is launched for that file, quit it.
4. Open another result by hitting the RETURN key.

Unexpected result:
The panel is cleared, showing just the '../' entry.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>